+++ docs/CHANGELOG.md
@@ -22,6 +22,7 @@
 
 * Farming mode won't try to move unmoveable reputation tokens.
 * Filters like stat:recovery:=0 now work (they couldn't match stat values of zero before).
+* Checking with VendorEngrams.xyz to see if 380 drops may be right for you.
 
 # 4.59.0 (2018-07-01)